Avoid rendering page 0 twice
Before we typically would render it twice because we do an explicit
update_window_textures() after creating the cache for page 0, but then
afterwards we will get an SDL_WINDOWEVENT for the window being shown.
This is fine but unnecessary.

Since it's risky to assume cross platform we will always get an
SDL_WINDOWEVENT for the first creation, we can't just remove the call to
update_window_textures(). Instead, initialise needs_redraw to 1 so that
on the first event loop we will render it.